refactor: move `parseContactPoint` to `SessionBuilder`
